+++ lib/ui/generate/generate_paper_wallet.dart
@@ -112,7 +112,8 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
       }
     });
     // Set initial currency format
-    _localCurrencyFormat = NumberFormat.currency(locale: widget.localCurrency.getLocale().toString(), symbol: widget.localCurrency.getCurrencySymbol());
+    _localCurrencyFormat = NumberFormat.currency(
+        locale: widget.localCurrency.getLocale().toString(), symbol: widget.localCurrency.getCurrencySymbol());
     // Set quick send amount
     if (quickSendAmount != null && quickSendAmount!.isNotEmpty && quickSendAmount != "0") {
       WidgetsBinding.instance.addPostFrameCallback((_) {
@@ -170,7 +171,8 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
       backgroundColor: StateContainer.of(context).curTheme.backgroundDark,
       body: LayoutBuilder(
         builder: (BuildContext context, BoxConstraints constraints) => SafeArea(
-          minimum: EdgeInsets.only(bottom: MediaQuery.of(context).size.height * 0.035, top: MediaQuery.of(context).size.height * 0.075),
+          minimum: EdgeInsets.only(
+              bottom: MediaQuery.of(context).size.height * 0.035, top: MediaQuery.of(context).size.height * 0.075),
           child: Column(
             children: <Widget>[
               // A widget that holds the header, the paragraph, the seed, "seed copied" text and the back button
@@ -238,7 +240,8 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
                       ),
                       // The paragraph
                       Container(
-                        margin: EdgeInsetsDirectional.only(start: smallScreen(context) ? 30 : 40, end: smallScreen(context) ? 30 : 40, top: 15.0),
+                        margin: EdgeInsetsDirectional.only(
+                            start: smallScreen(context) ? 30 : 40, end: smallScreen(context) ? 30 : 40, top: 15.0),
                         alignment: Alignment.centerLeft,
                         child: Column(
                           children: <Widget>[
@@ -283,7 +286,9 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
                               alignment: Alignment.topCenter,
                               children: <Widget>[
                                 Container(
-                                  margin: EdgeInsets.only(left: MediaQuery.of(context).size.width * 0.105, right: MediaQuery.of(context).size.width * 0.105),
+                                  margin: EdgeInsets.only(
+                                      left: MediaQuery.of(context).size.width * 0.105,
+                                      right: MediaQuery.of(context).size.width * 0.105),
                                   alignment: Alignment.bottomCenter,
                                   constraints: const BoxConstraints(maxHeight: 80, minHeight: 0),
                                 ),
@@ -319,8 +324,8 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                  AppButton.buildAppButton(context, AppButtonType.PRIMARY, AppLocalization.of(context)!.createGiftCard, Dimens.BUTTON_BOTTOM_DIMENS,
-                      onPressed: () {
+                  AppButton.buildAppButton(context, AppButtonType.PRIMARY, AppLocalization.of(context)!.createGiftCard,
+                      Dimens.BUTTON_BOTTOM_DIMENS, onPressed: () {
                     final bool isValid = _validateRequest();
                     if (!isValid) {
                       return;
@@ -359,17 +364,36 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
     return NumberUtil.truncateDecimal((valueLocal / conversion).toDecimal()).toString();
   }
 
-  String _convertCryptoToLocalCurrency() {
-    String convertedAmt = NumberUtil.sanitizeNumber(_sendAmountController!.text, maxDecimalDigits: 2);
-    if (convertedAmt.isEmpty) {
+  String _convertCryptoToLocalCurrency(String amount) {
+    String sanitizedAmt = amount
+        .replaceAll(_localCurrencyFormat!.symbols.GROUP_SEP, "")
+        .replaceAll(_localCurrencyFormat!.symbols.DECIMAL_SEP, ".");
+    sanitizedAmt = NumberUtil.sanitizeNumber(sanitizedAmt);
+    if (sanitizedAmt.isEmpty) {
       return "";
     }
-    final Decimal valueCrypto = Decimal.parse(convertedAmt);
+    final Decimal valueCrypto = Decimal.parse(sanitizedAmt);
     final Decimal conversion = Decimal.parse(StateContainer.of(context).wallet!.localCurrencyConversion!);
-    convertedAmt = NumberUtil.truncateDecimal(valueCrypto * conversion, digits: 2).toString();
-    convertedAmt = convertedAmt.replaceAll(".", _localCurrencyFormat!.symbols.DECIMAL_SEP);
-    convertedAmt = _localCurrencyFormat!.currencySymbol + convertedAmt;
-    return convertedAmt;
+    sanitizedAmt = NumberUtil.truncateDecimal(valueCrypto * conversion, digits: 2);
+
+    final List<String> splitStrs = sanitizedAmt.split(".");
+    final String firstPart = splitStrs[0].trim();
+    String secondPart = splitStrs.length > 1 ? splitStrs[1].trim() : "";
+
+    if (secondPart.isNotEmpty) {
+      secondPart = _localCurrencyFormat!.symbols.DECIMAL_SEP + secondPart;
+    }
+
+    final NumberFormat formatCurrency = NumberFormat.simpleCurrency(
+        decimalDigits: _localCurrencyFormat!.decimalDigits,
+        locale: _localCurrencyFormat!.locale,
+        name: _localCurrencyFormat!.currencyName);
+    String formattedCurrency = formatCurrency.format(int.parse(firstPart));
+    formattedCurrency = formattedCurrency
+            .substring(0, formattedCurrency.length - (_localCurrencyFormat!.decimalDigits! + 1))
+            .replaceAll(" ", "") +
+        secondPart;
+    return formattedCurrency;
   }
 
   // Determine if this is a max send or not by comparing balances
@@ -382,11 +406,12 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
       String textField = _sendAmountController!.text;
       String balance;
       if (_localCurrencyMode) {
-        balance = StateContainer.of(context)
-            .wallet!
-            .getLocalCurrencyBalance(context, StateContainer.of(context).curCurrency, locale: StateContainer.of(context).currencyLocale);
+        balance = StateContainer.of(context).wallet!.getLocalCurrencyBalance(
+            context, StateContainer.of(context).curCurrency,
+            locale: StateContainer.of(context).currencyLocale);
       } else {
-        balance = getRawAsThemeAwareFormattedAmount(context, StateContainer.of(context).wallet!.accountBalance.toString());
+        balance =
+            getRawAsThemeAwareFormattedAmount(context, StateContainer.of(context).wallet!.accountBalance.toString());
       }
       // Convert to Integer representations
       int textFieldInt;
@@ -398,12 +423,20 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
         balance = balance.replaceAll(_localCurrencyFormat!.symbols.GROUP_SEP, "");
         balance = balance.replaceAll(",", ".");
         final String sanitizedBalance = NumberUtil.sanitizeNumber(balance);
-        textFieldInt = (Decimal.parse(sanitizedTextField) * Decimal.fromInt(pow(10, NumberUtil.maxDecimalDigits) as int)).toDouble().toInt();
-        balanceInt = (Decimal.parse(sanitizedBalance) * Decimal.fromInt(pow(10, NumberUtil.maxDecimalDigits) as int)).toDouble().toInt();
+        textFieldInt =
+            (Decimal.parse(sanitizedTextField) * Decimal.fromInt(pow(10, NumberUtil.maxDecimalDigits) as int))
+                .toDouble()
+                .toInt();
+        balanceInt = (Decimal.parse(sanitizedBalance) * Decimal.fromInt(pow(10, NumberUtil.maxDecimalDigits) as int))
+            .toDouble()
+            .toInt();
       } else {
         textField = textField.replaceAll(",", "");
-        textFieldInt = (Decimal.parse(textField) * Decimal.fromInt(pow(10, NumberUtil.maxDecimalDigits) as int)).toDouble().toInt();
-        balanceInt = (Decimal.parse(balance) * Decimal.fromInt(pow(10, NumberUtil.maxDecimalDigits) as int)).toDouble().toInt();
+        textFieldInt = (Decimal.parse(textField) * Decimal.fromInt(pow(10, NumberUtil.maxDecimalDigits) as int))
+            .toDouble()
+            .toInt();
+        balanceInt =
+            (Decimal.parse(balance) * Decimal.fromInt(pow(10, NumberUtil.maxDecimalDigits) as int)).toDouble().toInt();
       }
       return textFieldInt == balanceInt;
     } catch (e) {
@@ -438,9 +471,12 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
       // Check our previous state
       if (_sendAmountController!.text == _lastCryptoAmount) {
         localAmountStr = _lastLocalCurrencyAmount;
+        if (!_lastLocalCurrencyAmount.startsWith(_localCurrencyFormat!.currencySymbol)) {
+          _lastLocalCurrencyAmount = _localCurrencyFormat!.currencySymbol + _lastLocalCurrencyAmount;
+        }
       } else {
         _lastCryptoAmount = _sendAmountController!.text;
-        _lastLocalCurrencyAmount = _convertCryptoToLocalCurrency();
+        _lastLocalCurrencyAmount = _convertCryptoToLocalCurrency(_sendAmountController!.text);
         localAmountStr = _lastLocalCurrencyAmount;
       }
       setState(() {
@@ -472,7 +508,9 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
               LengthLimitingTextInputFormatter(13),
               if (_localCurrencyMode)
                 CurrencyFormatter(
-                    decimalSeparator: _localCurrencyFormat!.symbols.DECIMAL_SEP, commaSeparator: _localCurrencyFormat!.symbols.GROUP_SEP, maxDecimalDigits: 2)
+                    decimalSeparator: _localCurrencyFormat!.symbols.DECIMAL_SEP,
+                    commaSeparator: _localCurrencyFormat!.symbols.GROUP_SEP,
+                    maxDecimalDigits: 2)
               else
                 CurrencyFormatter(maxDecimalDigits: NumberUtil.maxDecimalDigits),
               LocalCurrencyFormatter(active: _localCurrencyMode, currencyFormat: _localCurrencyFormat)
@@ -505,14 +543,16 @@ class _GeneratePaperWalletScreenState extends State<GeneratePaperWalletScreen> {
             return;
           }
           if (!_localCurrencyMode) {
-            _sendAmountController!.text = getRawAsThemeAwareAmount(context, StateContainer.of(context).wallet!.accountBalance.toString());
+            _sendAmountController!.text =
+                getRawAsThemeAwareAmount(context, StateContainer.of(context).wallet!.accountBalance.toString());
           } else {
-            String localAmount = StateContainer.of(context)
-                .wallet!
-                .getLocalCurrencyBalance(context, StateContainer.of(context).curCurrency, locale: StateContainer.of(context).currencyLocale);
+            String localAmount = StateContainer.of(context).wallet!.getLocalCurrencyBalance(
+                context, StateContainer.of(context).curCurrency,
+                locale: StateContainer.of(context).currencyLocale);
             localAmount = localAmount.replaceAll(_localCurrencyFormat!.symbols.GROUP_SEP, "");
             localAmount = localAmount.replaceAll(_localCurrencyFormat!.symbols.DECIMAL_SEP, ".");
-            localAmount = NumberUtil.sanitizeNumber(localAmount).replaceAll(".", _localCurrencyFormat!.symbols.DECIMAL_SEP);
+            localAmount =
+                NumberUtil.sanitizeNumber(localAmount).replaceAll(".", _localCurrencyFormat!.symbols.DECIMAL_SEP);
             _sendAmountController!.text = _localCurrencyFormat!.currencySymbol + localAmount;
           }
         },